+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - [#531](https://github.com/XenitAB/terraform-modules/pull/531) Make prefix configurable for Azure role definition names
+- [#524](https://github.com/XenitAB/terraform-modules/pull/524) Update grafana-agent to 0.1.5
 
 ### Fixed
 
@@ -21,7 +22,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - [#523](https://github.com/XenitAB/terraform-modules/pull/523) Update starboard to 0.14.0, only scan the latest deployments
   and set a TTL on the vulnerability reports to be recreated after 25 hours.
-- [#524](https://github.com/XenitAB/terraform-modules/pull/524) Update grafana-agent to 0.1.5
 
 ### Added